drm/bridge: synopsys: dw-mipi-dsi: Enter LP-11 by default on startup
This is a bit of a hack to fix the toshiba dp bridge chipset that
requires DSI to be in LP-11 when reset is deasserted. At the time of writing,
This is not solved in linux-drm, with various patches to have a host enter
LP-11 on request, or possibly have the pre_atomic always require LP-11. Until
this is solved more gracefully, we enable it on startup.